Moved auth* files to auth packages
